event: move middleware to the producer side in Stream (#114)
The new design replaces the application of middleware at the consumer,
which can be easily wrapped by the client side, with its application
at the producer.

The main implication of this change is that we can now guarantee that
the middleware will run against the caller context, not the async
derived context, which we need, e.g., to record events in
OpenTelemetry.